move negNat and subNat to Internal module
These functions were formerly private. They had to made accessible to other parts of the library to allow proofs, in particular for the IsLawfulNum instances. Both functions are still "hidden" in an "Internal" module though to signal these functions as not being intended to be used from any Haskell code.

This change also redefines subNat based on pattern-matching instead of
an if-else-expression with ltNat, to simplify proofs.