Prevent confusion from multiple anonymous modules
By using NameId instead of pretty-printed qualified name to determine if we are calling a function within the same module
we can distinguish between multiple anonymous records, as well as where-clauses.

This also prevents issues with modules where one module name happens to be a prefix of another.

Partially fixes issue #353